+++ lib/batch.dart
@@ -11,8 +11,12 @@ export 'package:batch/src/repeat_status.dart' show RepeatStatus;
 export 'package:batch/src/log_configuration.dart';
 export 'package:batch/src/log/logger_provider.dart';
 export 'package:batch/src/log/log_level.dart';
-export 'package:batch/src/log/log_input_event.dart';
-export 'package:batch/src/log/log_output_event.dart';
-export 'package:batch/src/log/log_filter.dart';
-export 'package:batch/src/log/log_printer.dart';
-export 'package:batch/src/log/log_output.dart';
+export 'package:batch/src/log/input_log_event.dart';
+export 'package:batch/src/log/output_log_event.dart';
+
+export 'package:batch/src/log/filter/log_filter.dart';
+export 'package:batch/src/log/filter/development_log_filter.dart';
+export 'package:batch/src/log/filter/production_log_filter.dart';
+
+export 'package:batch/src/log/output/console_log_output.dart';
+export 'package:batch/src/log/output/file_log_output.dart';