+++ lib/src/log/filter/log_filter.dart
@@ -2,12 +2,17 @@
 // Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:batch/src/log/log_input_event.dart';
+// Project imports:
+import 'package:batch/src/log/input_log_event.dart';
 import 'package:batch/src/log/log_level.dart';
 
 abstract class LogFilter {
   /// The log level
   LogLevel? level;
 
-  bool shouldLog(final LogInputEvent event);
+  void init() {}
+
+  bool shouldLog(final InputLogEvent event);
+
+  void dispose() {}
 }