+++ lib/src/log/filter/production_log_filter.dart
@@ -1,0 +1,14 @@
+// Copyright (c) 2022, Kato Shinya. All rights reserved.
+// Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// Project imports:
+import 'package:batch/src/log/filter/log_filter.dart';
+import 'package:batch/src/log/input_log_event.dart';
+
+/// Prints all logs with `level >= Logger.level` even in production.
+class ProductionLogFilter extends LogFilter {
+  @override
+  bool shouldLog(final InputLogEvent event) =>
+      event.level.index >= level!.index;
+}