+++ lib/src/log/printer/default_log_printer.dart
@@ -2,13 +2,14 @@
 // Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:batch/src/log/log_input_event.dart';
-import 'package:batch/src/log/log_printer.dart';
+// Project imports:
+import 'package:batch/src/log/input_log_event.dart';
+import 'package:batch/src/log/printer/log_printer.dart';
 
 /// The custom printer.
 class DefaultLogPrinter extends LogPrinter {
   @override
-  List<String> log(final LogInputEvent event) => [
+  List<String> log(final InputLogEvent event) => [
         '${DateTime.now()} [${event.level.name.padRight(7, ' ')}] :: ${event.message}'
       ];
 }