+++ lib/src/log/printer/log_printer.dart
@@ -2,10 +2,15 @@
 // Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:batch/src/log/log_input_event.dart';
+// Project imports:
+import 'package:batch/src/log/input_log_event.dart';
 
 abstract class LogPrinter {
+  void init() {}
+
   /// Is called every time a new [LogInputEvent] is sent and handles printing or
   /// storing the message.
-  List<String> log(final LogInputEvent event);
+  List<String> log(final InputLogEvent event);
+
+  void dispose() {}
 }